Reference

The following provides detailled documentation about types and methods provided by the InterpolationKernels package. This information is also available from the REPL by typing ? followed by the name of a method or a type.

Methods

InterpolationKernels.iscardinal — Function

iscardinal(ker)

yields whether the kernel ker is zero for non-zero integer arguments. Cardinal kernels are directly suitbale for interpolation.

InterpolationKernels.isnormalized — Function

isnormalized(ker)

yields whether the kernel ker has the partition of unity property. That is, the sum of the values computed by the kernel ker on a unit spaced grid is equal to one.

InterpolationKernels.brief — Function

brief(ker)

yields a brief description of the kernel ker.

InterpolationKernels.getweights — Function

getweights(ker, x - j) -> w1, w2, ..., wS

yields the S-tuple of the weights for interpolating around position x (in fractional index units) with kernel ker (S is the size of the kernel support). The index j is given by:

j = floor(x)   #  if `S` is even
j = round(x)   #  if `S` is odd

(the rounding direction does not matter for the result of the interpolation). The offset t = x - j is therefore in the range [0,1] for S even and in the range [-1/2,+1/2] for S odd.

Boundary conditions

InterpolationKernels.Boundaries — Type

All extrapolation methods (a.k.a. boundary conditions) are singletons and inherit from the abstract type Boundaries.

InterpolationKernels.Flat — Type

Type Flat indicates that flat boundary conditions hold. These conditions assume that the value of the nearest sample is used when extrapolating.

InterpolationKernels.SafeFlat — Type

Type SafeFlat indicates that flat boundary conditions hold. These conditions as similar to the ones assumed by Flat conditions but are safer because they account for possible integer overflows when converting coordinates to indices. As a consequence, applying SafeFlat boundary conditions is slower than applying Safe boundary conditions.

InterpolationKernels.boundaries — Function

boundaries(ker)

yields the type of the boundary conditions that hold for extrapolation with kernel ker.

Kernels

InterpolationKernels.Kernel — Type

Interpolation Kernels

An interpolation kernel Kernel{T,S,B} is parametrized by the floating-point type T of its coefficients, by the size S of its support and by the boundary conditions B to apply for extrapolation. For efficiency reasons, only kernels with (small) finite size supports are implemented.

A kernel ker is a callable object which may be used as a function with a real argument:

ker(x::Real)

yields kernel value at offset x. All kernel supports are symmetric; that is ker(x) is zero if abs(x) > S/2 with S the size of the kernel support.

Kernel conversion

The argument of a kernel be a floating-point type and/or a boundary conditions type:

ker(::Type{T}, ::Type{B}) where {T<:AbstractFloat, B<:Boundaries}

to convert the kernel to operate with given floating-point type T and use boundary conditions B (any of which can be omitted and their order is irrelevant). Beware that changing the floating-point type may lead to a loss of precision if the new floating-point type has more digits).

It is possible to change the floating-point type of a kernel or its boundary conditions by something like:

Float32(ker)    # change floating-point type of kernel `ker`
SafeFlat(ker)   # change boundary conditions of kernel `ker`

The above calls do not follow the usuall conventions that a constructor may be called to convert its argument(s) to an instance of its type, this is however practical considering the following more rigorous calls to repectively change the the floating-point type and boundary conditions of kernel ker:

convert(Kernel{Float32}, ker)
convert(Kernel{eltype(ker),length(ker),SafeFlat}, ker)

Available methods

The following methods are available for any interpolation kernel ker:

eltype(ker) -> T

yields the floating-point type for calculations,

length(ker) -> S
size(ker)   -> (S,)

yield the size the support of ker which is also the number of neighbors involved in an interpolation by this kernel,

boundaries(ker) -> B

yields the type of the boundary conditions applied for extrapolation; finally:

getweights(ker, t) -> w1, w2, ..., wS

yields the S interpolation weights for offset t

t = x - floor(x)        if s is even
    x - round(x)        if s is odd

t ∈ [0,1] if S is even or for offset t ∈ [-1/2,+1/2] if S is odd.

Rectangular B-spline

InterpolationKernels.RectangularSpline — Type

RectangularSpline([T=Float64,] B=Flat)

yields an instance of the rectangular spline for floating-point type T and boundary conditions B.

The rectangular spline (also known as box kernel, as Fourier window or as Dirichlet window) is the 1st order (constant) B-spline equals to 1 on [-1/2,+1/2), and 0 elsewhere.

The expression ker' yields a kernel instance which is the 1st derivative of the rectangular spline ker (also see the constructor RectangularSplinePrime).

InterpolationKernels.RectangularSplinePrime — Type

RectangularSplinePrime([T=Float64,] B=Flat)

yields a kernel instance that is the 1st derivative of the rectangular spline (see RectangularSpline) for floating-point type T and boundary conditions B.

Linear B-spline

InterpolationKernels.LinearSpline — Type

LinearSpline([T=Float64,] B=Flat)

yields an instance of the linear spline for floating-point type T and boundary conditions B.

The linear spline (also known as triangle kernel, as Bartlett window or as Fejér window) is the 2nd order (linear) B-spline given by:

ker(x) = 1 - |x|       if |x| ≤ 1
         0             if |x| ≥ 1

The expression ker' yields a kernel instance which is the 1st derivative of the linear spline ker (also see the constructor LinearSplinePrime).

InterpolationKernels.LinearSplinePrime — Type

LinearSplinePrime([T=Float64,] B=Flat)

yields a kernel instance that is the 1st derivative of the linear spline (see LinearSpline) for floating-point type T and boundary conditions B.

Quadratic B-spline

InterpolationKernels.QuadraticSpline — Type

QuadraticSpline([T=Float64,] B=Flat)

yields an instance of the quadratic spline for floating-point type T and boundary conditions B.

The quadratic spline is the 3rd order (quadratic) B-spline given by:

ker(x) = 3/4 - x^2                   if |x| ≤ 1/2
         (1/2)*(|x| - 3/2)^2         if |x| ≤ 3/2
         0                           if |x| ≥ 3/2

The expression ker' yields a kernel instance which is the 1st derivative of the quadratic spline ker (also see the constructor QuadraticSplinePrime).

InterpolationKernels.QuadraticSplinePrime — Type

QuadraticSplinePrime([T=Float64,] B=Flat)

yields a kernel instance that is the 1st derivative of the quadratic spline (see QuadraticSpline) for floating-point type T and boundary conditions B.

Cubic B-spline

InterpolationKernels.CubicSpline — Type

CubicSpline([T=Float64,] B=Flat)

yields an instance of the cubic spline for floating-point type T and boundary conditions B.

The 4th order (cubic) B-spline kernel is also known as Parzen window or as de la Vallée Poussin window. It is given by:

ker(x) = 2/3 + (|x|/2 - 1)*x^2       if |x| ≤ 1
         (1/6)*(2 - |x|)^2           if |x| ≤ 2
         0                           if |x| ≥ 2

The expression ker' yields a kernel instance which is the 1st derivative of the cubic spline ker (also see the constructor CubicSplinePrime).

InterpolationKernels.CubicSplinePrime — Type

CubicSplinePrime([T=Float64,] B=Flat)

yields a kernel instance that is the 1st derivative of the cubic spline (see CubicSpline) for floating-point type T and boundary conditions B.

Catmull-Rom interpolation kernel

InterpolationKernels.CatmullRomSpline — Type

CatmullRomSpline([T=Float64,] B=Flat)

yields an instance of the Catmull-Rom interpolation kernel for floating-point type T and boundary conditions B.

Catmull-Rom interpolation kernel is a piecewise cardinal cubic spline defined by:

ker(x) = ((3/2)*|x| - (5/2))*x^2 + 1             if |x| ≤ 1
         (((5/2) - (1/2)*|x|)*|x| - 4)*|x| + 2   if 1 ≤ |x| ≤ 2
         0                                       if |x| ≥ 2

Catmull-Rom kernel is a special case of Mitchell & Netravali kernel (see MitchellNetravaliSplinePrime).

The expression ker' yields a kernel instance which is the 1st derivative of the Catmull-Rom interpolation kernel ker (also see the constructor CatmullRomSplinePrime).

InterpolationKernels.CatmullRomSplinePrime — Type

CatmullRomSplinePrime([T=Float64,] B=Flat)

yields a kernel instance that is the 1st derivative of the Catmull-Rom interpolation kernel (see CatmullRomSpline) for floating-point type T and boundary conditions B.

The 1st derivative of the Catmull-Rom interpolation kernel is given by:

ker′(x) = ((9/2)*|x| - 5)*x                      if a = |x| ≤ 1
          (5 - (3/2)*|x|)*x - 4*sign(x)          if 1 ≤ |x| ≤ 2
          0                                      if |x| ≥ 2

Cardinal cubic splines

InterpolationKernels.CardinalCubicSpline — Type

CardinalCubicSpline([T=Float64,] c, B=Flat)

yields an instance of a cardinal cubic spline interpolation kernel for floating-point type T, tension parameter c and boundary conditions B.

The slope at x = ±1 is ±(c - 1)/2. Usually c ≤ 1, choosing c = 0 yields a Catmull-Rom spline (see CatmullRomSpline), c = 1 yields all zero tangents, c = -1 yields a truncated approximation of a cardinal sine, c = -1/2 yields an interpolating cubic spline with continuous second derivatives (inside its support).

The expression ker' yields a kernel instance which is the 1st derivative of the cardinal cubic spline ker (also see the constructor CardinalCubicSplinePrime).

InterpolationKernels.CardinalCubicSplinePrime — Type

CardinalCubicSplinePrime([T=Float64,] B=Flat)

yields a kernel instance that is the 1st derivative of the cardinal cubic spline interpolation kernel (see CardinalCubicSpline) for floating-point type T, tension parameter c and boundary conditions B.

Keys cardinal kernels

InterpolationKernels.KeysSpline — Type

KeysSpline([T=Float64,] a, B=Flat)

yields an instance of the Keys family of cardinal kernels for floating-point type T, parameter a and boundary conditions B.

These kernels are piecewise normalized cardinal cubic spline which depend on one parameter a which is the slope of the spline at abscissa 1. Keys cardinal kernels are defined by:

ker(x) = p(abs(x))   if abs(x) ≤ 1
         q(abs(x))   if 1 ≤ abs(x) ≤ 2
         0           if abs(x) ≥ 2

with:

p(x) = 1 - (a + 3)*x^2 + (a + 2)*x^3
q(x) = -4a + 8a*x - 5a*x^2 + a*x^3

The expression ker' yields a kernel instance which is the 1st derivative of the Keys kernel ker (also see the constructor KeysSplinePrime).

Reference:

Keys, Robert, G., "Cubic Convolution Interpolation for Digital Image Processing", IEEE Trans. Acoustics, Speech, and Signal Processing, Vol. ASSP-29, No. 6, December 1981, pp. 1153-1160.

InterpolationKernels.KeysSplinePrime — Type

KeysSplinePrime([T=Float64,] a, B=Flat)

yields a kernel instance that is the 1st derivative of the Keys kernel (see KeysSpline) for floating-point type T, parameter a and boundary conditions B. This derivative is given by:

ker′(x) = p′(abs(x))*sign(x)   if abs(x) ≤ 1
          q′(abs(x))*sign(x)   if 1 ≤ abs(x) ≤ 2
          0                    if abs(x) ≥ 2

with:

p(x) = -2*(a + 3)*x + 3*(a + 2)*x^2
q(x) = 8a - 10a*x + 3a*x^2

Mitchell-Netravali kernels

InterpolationKernels.MitchellNetravaliSpline — Type

MitchellNetravaliSpline([T=Float64,] [b=1/3, c=1/3,] B=Flat)

yields an instance of the Mitchell & Netravali family of kernels for floating-point type T, parameters b and c and boundary conditions B.

These kernels are cubic splines which depends on 2 parameters, b and c. Whatever the values of (b,c), all these kernels are normalized, even functions of class C¹ (these kernels and their first derivatives are continuous).

Taking b = 0 yields Keys's family of kernels and is a sufficient and necessary condition to have Mitchell & Netravali kernels be cardinal functions.

Using the constraint: b + 2c = 1 yields a cubic filter with, at least, quadratic order approximation.

Some specific values of (b,c) yield other well known kernels:

(b,c) = (1,0)     ==> cubic B-spline
(b,c) = (0,-a)    ==> Keys's cardinal cubics
(b,c) = (0,1/2)   ==> Catmull-Rom cubics
(b,c) = (b,0)     ==> Duff's tensioned B-spline
(b,c) = (1/3,1/3) ==> recommended by Mitchell-Netravali

The expression ker' yields a kernel instance which is the 1st derivative of the Mitchell & Netravali kernel ker (also see the constructor MitchellNetravaliSplinePrime).

Reference:

Mitchell & Netravali, "Reconstruction Filters in Computer Graphics", in Computer Graphics, Vol. 22, Num. 4 (1988). http://www.cs.utexas.edu/users/fussell/courses/cs384g/lectures/mitchell/Mitchell.pdf.

InterpolationKernels.MitchellNetravaliSplinePrime — Type

MitchellNetravaliSplinePrime([T=Float64,] [b=1/3, c=1/3,] B=Flat)

yields a kernel instance that is the 1st derivative of the Mitchell & Netravali kernel (see MitchellNetravaliSpline) for floating-point type T, parameters b and c and boundary conditions B.

Lanczos re-sampling kernels

InterpolationKernels.LanczosKernel — Type

LanczosKernel([T=Float64,] S, B=Flat)

yields an instance of a Lanczos re-sampling kernel of support size S (which must be even), for floating-point type T and boundary conditions B.

The Lanczos re-sampling kernels are even cardinal functions which tend to be normalized for large support size. They are defined by (see also https://en.wikipedia.org/wiki/Lanczos_resampling):

ker(x) = S/(2*(π*x)^2)*sin(π*x)*sin(2*π*x/S)     if |x| ≤ S/2
         0                                       if |x| ≥ S/2

The expression ker' yields the first derivative of a Lanczos re-sampling kernel ker (also see the constructor LanczosKernelPrime).

InterpolationKernels.LanczosKernelPrime — Type

LanczosKernelPrime([T=Float64,] S, B=Flat)

yields a kernel instance that is the 1st derivative of the Lanczos re-sampling kernel (see LanczosKernel) of support size S and for floating-point type T and boundary conditions B.